requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.1
coreapi==2.3.0
coreschema==0.0.4
coverage==4.0.3
cssselect==0.9.2
Django==1.9.6
django-cors-headers==2.0.2
django-crispy-forms==1.6.0
django-filter==0.15.3
django-guardian==1.4.5
django-rest-swagger==2.1.2
django-twitter-bootstrap==3.3.0
django-twitter-bootstrap-form==0.1.0
djangorestframework==3.6.2
djangorestframework-filters==0.9.1
djoser==0.5.2
drfdocs==0.0.11
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.1
itypes==1.1.0
jieba3k==0.35.1
Jinja2==2.9.5
lxml==3.6.4
MarkupSafe==1.0
mock==2.0.0
newspaper3k==0.1.7
nltk==3.2.1
openapi-codec==1.3.1
packaging==16.8
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.6.2
pyparsing==2.2.0
python-dateutil==2.5.3
PyYAML==3.12
requests==2.11.1
requests-file==1.4
selenium==2.53.6
simplejson==3.10.0
six==1.10.0
tldextract==2.0.1
tqdm==4.7.6
uritemplate==3.0.0
wordfencer==1.3.0
